requirements.txt
# WWAQ Parser Requirements
# Stand: 1. Tammus 5785

# Core
python-dateutil>=2.8.2
pyyaml>=6.0.1
requests>=2.31.0

# API
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.12.0
flake8>=7.0.0
python-dotenv>=1.0.0
